requirements.lib.txt
torch >= 1.8.0
torchvision >= 0.9.0
numpy >= 1.10
einops >= 0.1
ffmpeg-python >= 0.2.0
opencv-python >= 4.3
av >= 6.0
tqdm >= 4.0
pyyaml >= 5.0